# Use direct navigator for Backward Filter