--- a/benchmarks/cpu/toy_detector_cpu.cpp
+++ b/benchmarks/cpu/toy_detector_cpu.cpp
@@ -29,11 +29,15 @@
 
 // VecMem include(s).
 #include <vecmem/memory/host_memory_resource.hpp>
+#include <vecmem/utils/copy.hpp>
 
 // Google benchmark include(s).
 #include <benchmark/benchmark.h>
 
 BENCHMARK_DEFINE_F(ToyDetectorBenchmark, CPU)(benchmark::State& state) {
+
+    // VecMem copy object
+    vecmem::copy copy;
 
     // Type declarations
     using host_detector_type = traccc::default_detector::host;
@@ -54,7 +58,8 @@ BENCHMARK_DEFINE_F(ToyDetectorBenchmark, CPU)(benchmark::State& state) {
     traccc::host::track_params_estimation tp(host_mr);
     traccc::host::combinatorial_kalman_filter_algorithm host_finding(
         finding_cfg);
-    traccc::host::kalman_fitting_algorithm host_fitting(fitting_cfg, host_mr);
+    traccc::host::kalman_fitting_algorithm host_fitting(fitting_cfg, host_mr,
+                                                        copy);
 
     for (auto _ : state) {