--- a/core/include/traccc/fitting/fitting_config.hpp
+++ b/core/include/traccc/fitting/fitting_config.hpp
@@ -32,6 +32,10 @@ struct fitting_config {
     /// Smoothing with backward filter
     bool use_backward_filter = true;
     traccc::scalar covariance_inflation_factor = 1e3f;
+    std::size_t barcode_sequence_size_factor = 5;
+    std::size_t min_barcode_sequence_capacity = 100;
+    traccc::scalar backward_filter_mask_tolerance =
+        5.f * traccc::unit<scalar>::mm;
 };
 
 }  // namespace traccc