--- a/core/include/traccc/fitting/kalman_filter/gain_matrix_updater.hpp
+++ b/core/include/traccc/fitting/kalman_filter/gain_matrix_updater.hpp
@@ -95,6 +95,8 @@ struct gain_matrix_updater {
         trk_state.predicted().set_vector(predicted_vec);
         trk_state.predicted().set_covariance(predicted_cov);
 
+        // Flip the sign of projector matrix element in case the first element
+        // of line measurement is negative
         if constexpr (std::is_same_v<shape_t, detray::line<true>> ||
                       std::is_same_v<shape_t, detray::line<false>>) {