--- a/core/src/fitting/kalman_fitting_algorithm.cpp
+++ b/core/src/fitting/kalman_fitting_algorithm.cpp
@@ -11,8 +11,8 @@
 namespace traccc::host {
 
 kalman_fitting_algorithm::kalman_fitting_algorithm(
-    const config_type& config, vecmem::memory_resource& mr,
+    const config_type& config, vecmem::memory_resource& mr, vecmem::copy& copy,
     std::unique_ptr<const Logger> logger)
-    : messaging(std::move(logger)), m_config{config}, m_mr{mr} {}
+    : messaging(std::move(logger)), m_config{config}, m_mr{mr}, m_copy(copy) {}
 
 }  // namespace traccc::host