--- a/core/src/fitting/kalman_fitting_algorithm_constant_field_default_detector.cpp
+++ b/core/src/fitting/kalman_fitting_algorithm_constant_field_default_detector.cpp
@@ -33,7 +33,8 @@ kalman_fitting_algorithm::output_type kalman_fitting_algorithm::operator()(
         fitter{det, field, m_config};
 
     // Perform the track fitting using a common, templated function.
-    return details::fit_tracks(fitter, track_candidates, m_mr.get());
+    return details::fit_tracks(fitter, track_candidates, m_mr.get(),
+                               m_copy.get());
 }
 
 }  // namespace traccc::host