--- a/examples/options/src/track_fitting.cpp
+++ b/examples/options/src/track_fitting.cpp
@@ -38,6 +38,21 @@ track_fitting::track_fitting() : interface("Track Fitting Options") {
         po::value(&m_config.covariance_inflation_factor)
             ->default_value(m_config.covariance_inflation_factor),
         "Covariance inflation factor for the track fit");
+    m_desc.add_options()(
+        "barcode-sequence-size-factor",
+        po::value(&m_config.barcode_sequence_size_factor)
+            ->default_value(m_config.barcode_sequence_size_factor),
+        "Size factor for the barcode sequence used in the backward filter");
+    m_desc.add_options()(
+        "min-barcode-sequence-capacity",
+        po::value(&m_config.min_barcode_sequence_capacity)
+            ->default_value(m_config.min_barcode_sequence_capacity),
+        "Minimum capacity of barcode sequence");
+    m_desc.add_options()(
+        "backward-filter-mask-tolerance",
+        po::value(&m_config.backward_filter_mask_tolerance)
+            ->default_value(m_config.backward_filter_mask_tolerance),
+        "Mask tolerance for the backward filter");
 }
 
 track_fitting::operator fitting_config() const {
@@ -63,6 +78,15 @@ std::unique_ptr<configuration_printable> track_fitting::as_printable() const {
     cat->add_child(std::make_unique<configuration_kv_pair>(
         "Covariance inflation factor",
         std::to_string(m_config.covariance_inflation_factor)));
+    cat->add_child(std::make_unique<configuration_kv_pair>(
+        "Barcode sequence size factor",
+        std::to_string(m_config.barcode_sequence_size_factor)));
+    cat->add_child(std::make_unique<configuration_kv_pair>(
+        "Minimum capacity of barcode sequence",
+        std::to_string(m_config.min_barcode_sequence_capacity)));
+    cat->add_child(std::make_unique<configuration_kv_pair>(
+        "Mask tolerance for the backward filter",
+        std::to_string(m_config.backward_filter_mask_tolerance)));
 
     return cat;
 }