## PR description

The PR depends on acts-project/detray#955. I will make a new detray tag once that PR is merged

This PR replaces the normal navigator with direct navigator in the backward filter. In the direct navigator, only the sequence of surfaces given as an input argument can be found during the geometry navigation. The sequence is provided by the forward filter so that both filters can go thru the same set of surfaces. The navigation logic is much simpler, thus, the computing performance is supposed to be improved. Another advantage of the direct navigator is that a relatively large mask tolerance can be used as there is no geometry overlapping issue with multiple surfaces. This means that the backward filter will not miss the surfaces if we set the mask tolerance properly. Even the test for wire chamber, which has notoriously small cross section (2-3 cm) for each wire, passes without using the RTS smoother -- We will come back to this in physical performance section. 

(I had to add `vecmem::copy` object to host fitter class to use the direct navigator, which is responsible for the most of file changes.)

All data or plots are from single precision

## Computing Performance

### ODD Benchmark

Summary: 
Track fitting time 6074 ms => 4327 ms

```
./bin/traccc_seq_example_cuda \
--digitization-file=geometries/odd/odd-digi-geometric-config.json \
--detector-file=geometries/odd/odd-detray_geometry_detray.json \
--material-file=geometries/odd/odd-detray_material_detray.json \
--grid-file=geometries/odd/odd-detray_surface_grids_detray.json \
--input-directory=odd/odd-simulations-20240509/geant4_ttbar_mu200/ \
--use-detray-detector=true \
--input-events=10
```

**This PR**

```
19:13:13    CudaSeqExample                INFO      ==> Statistics ... 
19:13:13    CudaSeqExample                INFO      - read    3790687 cells
19:13:13    CudaSeqExample                INFO      - created (cpu)  0 measurements     
19:13:13    CudaSeqExample                INFO      - created (cuda)  1042450 measurements     
19:13:13    CudaSeqExample                INFO      - created (cpu)  0 spacepoints     
19:13:13    CudaSeqExample                INFO      - created (cuda) 823334 spacepoints     
19:13:13    CudaSeqExample                INFO      - created  (cpu) 0 seeds
19:13:13    CudaSeqExample                INFO      - created (cuda) 368854 seeds
19:13:13    CudaSeqExample                INFO      - found (cpu)    0 tracks
19:13:13    CudaSeqExample                INFO      - found (cuda)   918894 tracks
19:13:13    CudaSeqExample                INFO      - fitted (cpu)   0 tracks
19:13:13    CudaSeqExample                INFO      - fitted (cuda)  918894 tracks
19:13:13    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  8068 ms
19:13:13    CudaSeqExample                INFO               Clusterization (cuda)  33 ms
19:13:13    CudaSeqExample                INFO         Spacepoint formation (cuda)  4 ms
19:13:13    CudaSeqExample                INFO                      Seeding (cuda)  378 ms
19:13:13    CudaSeqExample                INFO                 Track params (cuda)  4 ms
19:13:13    CudaSeqExample                INFO                Track finding (cuda)  3042 ms
19:13:13    CudaSeqExample                INFO                Track fitting (cuda)  4327 ms
19:13:13    CudaSeqExample                INFO                           Wall time  15949 ms
```

**Main**

```
19:14:18    CudaSeqExample                INFO      ==> Statistics ... 
19:14:18    CudaSeqExample                INFO      - read    3790687 cells
19:14:18    CudaSeqExample                INFO      - created (cpu)  0 measurements     
19:14:18    CudaSeqExample                INFO      - created (cuda)  1042450 measurements     
19:14:18    CudaSeqExample                INFO      - created (cpu)  0 spacepoints     
19:14:18    CudaSeqExample                INFO      - created (cuda) 823334 spacepoints     
19:14:18    CudaSeqExample                INFO      - created  (cpu) 0 seeds
19:14:18    CudaSeqExample                INFO      - created (cuda) 368854 seeds
19:14:18    CudaSeqExample                INFO      - found (cpu)    0 tracks
19:14:18    CudaSeqExample                INFO      - found (cuda)   918996 tracks
19:14:18    CudaSeqExample                INFO      - fitted (cpu)   0 tracks
19:14:18    CudaSeqExample                INFO      - fitted (cuda)  918996 tracks
19:14:18    CudaSeqExample                INFO      ==>Elapsed times...            File reading  (cpu)  8045 ms
19:14:18    CudaSeqExample                INFO               Clusterization (cuda)  35 ms
19:14:18    CudaSeqExample                INFO         Spacepoint formation (cuda)  4 ms
19:14:18    CudaSeqExample                INFO                      Seeding (cuda)  378 ms
19:14:18    CudaSeqExample                INFO                 Track params (cuda)  4 ms
19:14:18    CudaSeqExample                INFO                Track finding (cuda)  2996 ms
19:14:18    CudaSeqExample                INFO                Track fitting (cuda)  6074 ms
19:14:18    CudaSeqExample                INFO                           Wall time  17559 ms
```

### Continuous Benchmark with Toy Detector

Unfortunately, I don't see a noticeable change in the toy detector benchmark. Strictly speaking it decreases a little. However, it should be noted that the backward filter of the `main` branch may exit quite early due to the unstable navigation while the backward filter with direct navigator mostly survives by the end of surface sequence. 

Edit: Another reason might be that the fitter is using minimal search window of {0,0} so the navigation cost was already pretty small

**This PR**

```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
ToyDetectorBenchmark/CUDA/real_time 7193191335 ns   7192873293 ns            1 event_throughput_Hz=13.902/s
```

**Main**

```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
ToyDetectorBenchmark/CUDA/real_time 7175976105 ns   7175235093 ns            1 event_throughput_Hz=13.9354/s
```

## Physical Performance


### Truth fitting in ODD

I ran the truth fitting on ODD with simplified muon simulation. (See the bash script below) 

<details>

<summary>Bash script</summary>

```sh
#!/bin/bash                                                                                                        
n_ptc=1000                                                                                           
mom_array=(1 10 100)                                                         
mask_tol_array=(0 0.1 0.5 1.0 2 3 4 5)

for mom in "${mom_array[@]}"
do

    command_simulate="                                                                                             
    ./traccc_build/bin/traccc_simulate \                                                                           
    --detector-file=geometries/odd/odd-detray_geometry_detray.json \                                               
    --material-file=geometries/odd/odd-detray_material_detray.json \                                               
    --grid-file=geometries/odd/odd-detray_surface_grids_detray.json \                                              
    --output-directory=odd/simplified_${n_ptc}muon_${mom}GeV/ \                                                    
    --gen-events=1 \                                                                                               
    --gen-nparticles=${n_ptc} \                                                                                    
    --gen-mom-gev=${mom}:${mom} \                                                                                  
    --gen-eta=-1:1 \                                                                                               
    --rk-tolerance=1e-8                                                                                            
    "

    ${command_simulate}

    for tol in "${mask_tol_array[@]}"
    do

        echo "Mom setup: ${mom}, Tol setup: ${tol}"

        command_truth_fit="                                                                                        
        ./traccc_build/bin/traccc_truth_fitting_example \                                                          
        --detector-file=geometries/odd/odd-detray_geometry_detray.json \                                           
        --material-file=geometries/odd/odd-detray_material_detray.json \                                           
        --grid-file=geometries/odd/odd-detray_surface_grids_detray.json \                                          
        --use-detray-detector=true \                                                                               
        --use-acts-geom-source=false \                                                                             
        --input-directory=odd/simplified_${n_ptc}muon_${mom}GeV/ \                                                 
        --input-events=1 \                                                                                         
        --backward-filter-mask-tolerance=${tol} \                                                                  
        --check-performance                                                                                        
        "
        #--min-mask-tolerance=${tol} \                                                                             
        #--max-mask-tolerance=${tol} \                                                                             
        #--overstep-tolerance-um=${tol} \                                                                          

        ${command_truth_fit}

        mv ./performance_track_fitting.root ./traccc_build/performance_track_fitting_${mom}_${tol}.root
    done
done
```

</details>


Following figures show the number of successfully fitted track, tracks with non-positive NDF, and tracks whose all tracks states are not smoothed, as a function of mask tolerance for the backward filter (Note that 1 or 2 tracks out of 1000 tracks make an error, thus, not included in the plot ). As expected, the number of tracks with imperfect smoothing drops almost to zero with large mask tolerance. The non-postive NDF tracks, which is from the forward filter, is not supposed to happen if the forward filter follows the same track of CKF propagation... this is under investigation


<img src="https://github.com/user-attachments/assets/d595daf5-0798-4465-bc37-320955fab04a" width="400" />


What is really unfortunate, the pull value or p-value distribution of **1 GeV/c** tracks do not look good, which will be studied later. The concentration around 0 and 1 of p-value distribution is crazy:

#### 1 GeV/c with material

<img src="https://github.com/user-attachments/assets/cfa370c9-89b0-4712-a69e-3303748890cf" width="300" />
<img src="https://github.com/user-attachments/assets/edd53e16-5828-4bfd-8e0c-e2a3f3cd0956" width="300" />

Material interaction with low momentum is definitely an issue.

#### 1 GeV/c without material

<img src="https://github.com/user-attachments/assets/51e96cb2-6f93-4f70-a9ef-32cafd730b98" width="300" />
<img src="https://github.com/user-attachments/assets/90b3e9f4-347c-4000-8cf1-5220359a1428" width="300" />

### Truth fitting in Wire Chamber

As mentioned above, the wire chamber test pass with the backward filter and reasonable size of initial covariance. However, the p-value test function is commented as its distribution is concentrated at zero:

<img src="https://github.com/user-attachments/assets/a3c03039-4ce8-4083-a154-1759a509b90b" width="300" />

Not surprising at all as tracks with wrong left-right hit assignment might result into low p-values, but I still need to investigate in detail. Anyway It makes much more sense than the ODD p-value distribution above.