@stephenswat 2025-03-18
Nice! I would expect this to be a bit faster than the normal navigator; is that right? Do you have any performance numbers?